mlmodel_flask/requirements.txt
flask
flask-cors
fastai
Pillow
google-generativeai